chapter04: Counting Sort and Really Simple Sort (Let's just use Int for these)

+++ chapter04/reallysimplesort.go
@@ -9,7 +9,7 @@ package chapter04
 // and we don't need to care about satellite data.
 // We can see that the algorithm would take only Theta(n) time instead of
 // Theta(n * lg n).
-func ReallySimpleSort(A []float64, n int) {
+func ReallySimpleSort(A []int, n int) {
 
 	k := 0
